added passport for security issues, added webpack hmr, integrate mongodb in-memory server for running tests, add auth/register route for creating new users in database

+++ src/todos/todos.service.ts
@@ -2,8 +2,8 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model, Query } from 'mongoose';
-import { Todo, TodoModel } from './schemas/todo.schema';
+import { Model, Query, Types } from 'mongoose';
+import { Todo, TodoDocument, TodoModel } from './schemas/todo.schema';
 import { QueryParams } from './todos.interfaces';
 
 @Injectable()
@@ -23,35 +23,43 @@ export class TodosService {
     return await this.todoModel.create(createTodoDto);
   }
 
-  async findAll(params: QueryParams) {
+  async findAll(
+    user_id: Types.ObjectId,
+    params: QueryParams,
+  ): Promise<TodoDocument[]> {
     console.log(`This action returns all todos`);
-    let todos = [];
-    if (params.sort) {
-      todos = await this.todoModel
-        .find()
-        .sort({ [params.sort_by]: params.sort });
-    } else {
-      todos = await this.todoModel.find();
+    let sort = {};
+    if (params.sort && params.sort_by) {
+      sort = { [params.sort_by]: params.sort };
     }
+    const todos = await this.todoModel.find({ user_id: user_id }).sort(sort);
     if (params.filter) {
       return todos.filter((todo) => todo[params.filter] === params.filter_by);
     }
     return todos;
   }
 
-  async findOne(id: string) {
+  async findOne(user_id: Types.ObjectId, id: string) {
     console.log(`This action returns a #${id} todo`);
-    const existing_todo = await this.todoModel.findById(id);
+    const existing_todo = await this.todoModel.findOne({
+      _id: id,
+      user_id: user_id,
+    });
     if (existing_todo) {
       return existing_todo;
     }
     throw new HttpException('Todo was not found.', HttpStatus.NOT_FOUND);
   }
 
-  async update(id: string, updateTodoDto: UpdateTodoDto) {
+  async update(
+    user_id: Types.ObjectId,
+    id: string,
+    updateTodoDto: UpdateTodoDto,
+  ) {
     console.log(`This action updates a #${id} todo`);
     return await this.todoModel.updateOne(
       {
+        user_id: user_id,
         _id: id,
       },
       {
@@ -61,11 +69,12 @@ export class TodosService {
   }
 
   async remove(
+    user_id: Types.ObjectId,
     id: string,
   ): Promise<
     Query<{ ok?: number; n?: number; deletedCount?: number }, TodoModel>
   > {
     console.log(`This action removes a #${id} todo`);
-    return await this.todoModel.deleteOne({ _id: id });
+    return await this.todoModel.deleteOne({ user_id: user_id, _id: id });
   }
 }
